LayoutTests/platform/win/fast/forms/input-maxlength-ime-completed-expected.txt

CONSOLE MESSAGE: line 7: ReferenceError: Can't find variable: textInputController
There was a bug that users could input text longer than maxlength via IME. This test confirms it was fixed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
